+++ apps/api/src/common/interceptors/mentor-documents.interceptor.spec.ts
@@ -1,0 +1,7 @@
+import { MentorDocumentsInterceptor } from './mentor-documents.interceptor';
+
+describe('MentorDocumentsInterceptor', () => {
+  it('should be defined', () => {
+    expect(new MentorDocumentsInterceptor()).toBeDefined();
+  });
+});
